make webidentity/clientgrants response classes as inner-classes (#1040)
